refactor(react-query): remove unnecessary test file
